[Devesh]: Added ability to calculate total fare for multiple rides

+++ src/main/java/com/cabinvoicegenerator/InvoiceData.java
@@ -2,10 +2,10 @@ package com.cabinvoicegenerator;
 
 public class InvoiceData {
 	
-	public Integer kiloMeters;
+	public Double kiloMeters;
 	public Integer time;
 	
-	public InvoiceData(Integer kiloMeters,Integer time) {
+	public InvoiceData(Double kiloMeters,Integer time) {
 		this.kiloMeters = kiloMeters;
 		this.time = time;
 	}
